llm-app/backend/requirements.txt
debugpy==1.8.13
fastapi==0.115.11
fastapi-camelcase==2.0.0
firebase_admin==6.7.0
pyhumps==3.8.0
mollie-api-python==3.7.4
mongoengine==0.29.1
pydantic==2.10.6
python-dotenv==1.0.1
starlette==0.46.1
uvicorn==0.34.0
